chore: move constants and files
