+++ packages/styled/src/builders/border.ts
@@ -1,7 +1,7 @@
 import { CharcoalAbstractTheme } from '@charcoal-ui/theme'
 import { CSSObject } from 'styled-components'
 import { objectKeys } from '../util'
-import { Internal, createInternal } from './internal'
+import { Internal, createInternal } from '../internals'
 import {
   defineConstantProperties,
   defineProperties,
