+++ packages/styled/src/builders/borderRadius.ts
@@ -1,6 +1,6 @@
 import { CharcoalAbstractTheme } from '@charcoal-ui/theme'
 import { px } from '@charcoal-ui/utils'
-import { Internal, createInternal } from './internal'
+import { Internal, createInternal } from '../internals'
 import { defineConstantProperties } from '../factories/lib'
 
 export const createBorderRadiusCss =
